+++ src/combinator/ops/mul/inline.rs
@@ -1,4 +1,4 @@
-use super::{impl_mul, impl_mul_with_sep, Mul, Repeat, Sep};
+use super::{impl_mul, Mul, Repeat};
 use crate::{
   action::{shift_input, Action, Input, Output},
   combinator::Combinator,
@@ -9,9 +9,9 @@ use std::ops;
 /// See [`Combinator::fold`].
 #[derive(Copy, Clone)]
 pub struct InlineFold<T, Init, Folder> {
-  action: T,
-  init: Init,
-  fold: Folder,
+  pub(super) action: T,
+  pub(super) init: Init,
+  pub(super) fold: Folder,
 }
 
 impl<T: fmt::Debug, Init, Folder> fmt::Debug for InlineFold<T, Init, Folder> {
@@ -78,38 +78,6 @@ unsafe impl<
   }
 }
 
-impl<T, S, Acc, Repeater: Repeat, Initializer: Fn() -> Acc, InlineFolder>
-  ops::Mul<(Repeater, Initializer, InlineFolder)> for Sep<T, S>
-{
-  type Output = Combinator<Mul<Sep<T, S>, (Repeater, Initializer, InlineFolder)>>;
-
-  /// See [`ops::mul`](crate::combinator::ops::mul) for more information.
-  #[inline]
-  fn mul(self, rhs: (Repeater, Initializer, InlineFolder)) -> Self::Output {
-    Self::Output::new(Mul::new(self, rhs))
-  }
-}
-
-unsafe impl<
-    State,
-    Heap,
-    T: Action<State, Heap>,
-    S: Action<State, Heap>,
-    Acc,
-    Repeater: Repeat,
-    Initializer: Fn() -> Acc,
-    InlineFolder: Fn(T::Value, Acc, Input<&mut State, &mut Heap>) -> Acc,
-  > Action<State, Heap> for Mul<Sep<T, S>, (Repeater, Initializer, InlineFolder)>
-{
-  type Value = Acc;
-
-  #[inline]
-  fn exec(&self, mut input: Input<&mut State, &mut Heap>) -> Option<Output<Self::Value>> {
-    let (repeat, init, fold) = &self.rhs;
-    impl_mul_with_sep!(input, repeat, init, fold, self.lhs.value, self.lhs.sep)
-  }
-}
-
 #[cfg(test)]
 mod tests {
   use super::*;
@@ -127,12 +95,7 @@ mod tests {
 
   #[test]
   fn test_inline_fold_with_sep() {
-    let combinator = eat('a').bind(1).sep(',')
-      * (
-        1..,
-        || 0,
-        |v, acc, _: Input<&mut (), &mut ()>| acc + v | acc + v,
-      );
+    let combinator = (eat('a').bind(1).fold(|| 0, |v, acc, _| acc + v) * (1..)).sep(',');
     let output = combinator
       .exec(Input::new(Instant::new("a,a,a"), &mut (), &mut ()).unwrap())
       .unwrap();